moved typing to utils
