+++ mrmustard/lab/circuit.py
@@ -25,7 +25,7 @@ from typing import List, Optional, Tuple
 from mrmustard import settings
 from mrmustard.lab.abstract import State, Transformation
 from mrmustard.training import Parametrized
-from mrmustard.typing import RealMatrix, RealVector
+from mrmustard.utils.typing import RealMatrix, RealVector
 from mrmustard.utils.circdrawer import circuit_text
 from mrmustard.utils.xptensor import XPMatrix, XPVector
