+++ mrmustard/math/lattice/strategies/beamsplitter.py
@@ -26,7 +26,7 @@ import numpy as np
 from numba import njit
 
 from mrmustard.math.lattice import steps
-from mrmustard.typing import ComplexMatrix, ComplexTensor, ComplexVector
+from mrmustard.utils.typing import ComplexMatrix, ComplexTensor, ComplexVector
 
 SQRT = np.sqrt(np.arange(100000))
