+++ mrmustard/physics/gaussian.py
@@ -22,7 +22,7 @@ from thewalrus.quantum import is_pure_cov
 
 from mrmustard import settings
 from mrmustard.math import Math
-from mrmustard.typing import Matrix, Scalar, Vector
+from mrmustard.utils.typing import Matrix, Scalar, Vector
 from mrmustard.utils.xptensor import XPMatrix, XPVector
 
 math = Math()
